+++ src/main/java/com/interviewai/backend/auth/service/AuthServiceImpl.java
@@ -2,8 +2,8 @@ package com.interviewai.backend.auth.service;
 
 import com.interviewai.backend.auth.enums.AuthErrorCode;
 import com.interviewai.backend.auth.service.dto.AuthTokenServiceDto;
-import com.interviewai.backend.common.exception.BusinessException;
-import com.interviewai.backend.config.jwt.JwtProvider;
+import com.interviewai.backend.global.common.exception.BusinessException;
+import com.interviewai.backend.global.config.jwt.JwtProvider;
 import com.interviewai.backend.user.enums.UserErrorCode;
 import com.interviewai.backend.user.model.User;
 import com.interviewai.backend.user.repository.UserRepository;
