+++ src/main/java/com/interviewai/backend/global/common/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.interviewai.backend.common.exception;
+package com.interviewai.backend.global.common.exception;
 
 import org.springframework.http.HttpStatus;
