+++ src/main/java/com/interviewai/backend/global/config/oauth/GithubOAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package com.interviewai.backend.config.oauth;
+package com.interviewai.backend.global.config.oauth;
 
 import java.util.Map;
